Implement limited action token functionality
This commit adds a limited action token system where users are allotted a certain number of tokens to spend on actions. Two new properties were added to the User model: 'tokenCount' as an integer and 'lastTokenReset' as a timestamp. The 'pay_action_cost' decorator was introduced to manage token deduction for specific routes. New tests have been added to ensure correct functionality of the action token system. Furthermore, the README has been updated to reflect this change.