Remove pytest skip marks from test functions
This commit simply removes the skip decorators from several test functions in the test_app.py file, indicating these tests are now ready to be included in the test run. Additionally, the "dev" mark has been added to a couple of test functions, possibly indicating them as under development or for use in a development environment.